feature:record 모듈의 View 기반 코드 삭제 (#133)
* #132 - remove : record의 view 기반 코드 삭제

* #132 - style : record:screen 밑에 있는 파일들을 record 패키지 밑으로 이동
