+++ feature/main/src/main/java/com/example/main/app/RoutinerApp.kt
@@ -17,7 +17,7 @@ import com.example.main.screen.navigation.MAIN_ROUTE
 import com.example.main.screen.navigation.mainNavGraph
 import com.hegunhee.daily.navigation.DAILY_ROUTE
 import com.hegunhee.daily.navigation.dailyNavGraph
-import com.hegunhee.record.screen.navigation.recordNavGraph
+import com.hegunhee.record.navigation.recordNavGraph
 import com.hegunhee.repeat.screen.navigation.repeatRoutineNavGraph
 import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_REPEAT_ROUTINE_ROUTE
 import com.hegunhee.routiner.insertRoutine.screen.navigation.INSERT_ROUTINE_ROUTE
