+++ feature/main/src/main/java/com/example/main/drawer/DrawerItem.kt
@@ -3,7 +3,7 @@ package com.example.main.drawer
 import androidx.annotation.DrawableRes
 import com.example.main.R
 import com.hegunhee.daily.navigation.DAILY_ROUTE
-import com.hegunhee.record.screen.navigation.RECORD_ROUTE
+import com.hegunhee.record.navigation.RECORD_ROUTE
 import com.hegunhee.repeat.screen.navigation.REPEAT_ROUTINE_ROUTE
 import com.hegunhee.setting.screen.navigation.SETTING_ROUTE
