--- feature/record/src/main/java/com/hegunhee/record/dateSelector/DateSelectorActionHandler.kt
@@ -1,8 +1,0 @@
-package com.hegunhee.record.dateSelector
-
-import hegunhee.routiner.model.Date
-
-interface DateSelectorActionHandler {
-
-    fun onDateSelectorClick(date : Date)
-}
