--- feature/record/src/main/java/com/hegunhee/record/screen/ReviewState.kt
@@ -1,17 +1,0 @@
-package com.hegunhee.record.screen
-
-import hegunhee.routiner.model.Review
-
-sealed interface ReviewState {
-
-    data object Loading : ReviewState
-
-    data class Exist(
-        val review: Review,
-    ) : ReviewState
-
-    data object Revise : ReviewState
-
-    data object Empty : ReviewState
-
-}
