--- feature/record/src/main/res/layout/item_date_selector.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_margin="10dp">
-
-    <TextView
-        android:id="@+id/yearDesc"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/white"
-        android:textSize="13sp"
-        app:layout_constraintEnd_toEndOf="@id/dayDesc"
-        app:layout_constraintStart_toStartOf="@id/dayDesc"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="2024" />
-
-    <TextView
-        android:id="@+id/dayDesc"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:textColor="@color/white"
-        android:textSize="11sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/yearDesc"
-        tools:text="3/28 (목)" />
-
-    <ImageView
-        android:id="@+id/selectDot"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="5dp"
-        android:background="@drawable/ic_circle_12"
-        android:backgroundTint="@color/white"
-        android:visibility="invisible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="@id/dayDesc"
-        app:layout_constraintStart_toStartOf="@id/dayDesc"
-        app:layout_constraintTop_toBottomOf="@id/dayDesc" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
