--- feature/record/src/main/res/layout/record_item.xml
@@ -1,68 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout>
-
-    <data>
-
-        <variable
-            name="routine"
-            type="hegunhee.routiner.model.Routine" />
-
-        <import type="android.view.View" />
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/rounded_shape_gray_20dp"
-        android:layout_margin="10dp" >
-
-        <ImageView
-            android:id="@+id/finish_image"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:src="@drawable/ic_check"
-            android:visibility="@{routine.finished ? View.VISIBLE : View.GONE}"
-            app:layout_constraintBottom_toBottomOf="@id/routine_title"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="@id/routine_title" />
-
-        <TextView
-            android:id="@+id/routine_title"
-            android:layout_width="200dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:layout_marginTop="10dp"
-            android:layout_marginBottom="10dp"
-            android:ellipsize="end"
-            android:maxLines="1"
-            android:text="@{routine.text}"
-            android:textColor="@color/white"
-            android:textSize="30sp"
-            android:textStyle="bold"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toEndOf="@id/finish_image"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="영어 공부하기" />
-
-        <TextView
-            android:id="@+id/category"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="20dp"
-            android:layout_marginStart="8dp"
-            android:text="@{routine.category}"
-            android:textColor="@color/white"
-            android:textSize="20sp"
-            android:visibility="@{routine.category.length() == 0 ? View.INVISIBLE : View.VISIBLE}"
-            app:layout_constraintBottom_toBottomOf="@id/routine_title"
-            app:layout_constraintTop_toTopOf="@id/routine_title"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:backgroundTint="@color/gray"
-            android:background="@drawable/rounded_shape_gray_20dp"
-            tools:text="영어" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
